(VR) Began working on the inventory system
