--- Assets/Scripts/Item/Inventory.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 55bec35914633e849844346111fa4f50
