+++ Assets/Scripts/Item/InventoryVR.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f5850fe45a2ff9849807abe3cbade14f
