+++ Assets/Scripts/Item/InventoryXR.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 55bec35914633e849844346111fa4f50
